refactor(activity-view): split scope.py windows out into windows.py
``activity/scope.py`` mixed pure interval arithmetic
(``_intersect_windows`` / ``_union_windows`` / ``_merge_entity_windows`` /
``_row_within_any_window``) with DB-touching scope resolution
(``_resolve_dashboard_scope`` / ``_resolve_chart_scope``). The pure
functions were called by ``activity/queries._fetch_change_records``
via a lazy import to dodge a ``scope ↔ queries`` cycle.

Pull the four pure functions into ``activity/windows.py``:
* ``scope.py`` keeps the four DB-touching resolvers and imports the
  pure helpers from ``windows.py``.
* ``queries.py`` now imports ``_row_within_any_window`` from
  ``windows.py`` at module-top; the lazy import inside
  ``_fetch_change_records`` is dropped.
* ``windows.py`` depends only on ``kinds.py`` for the ``Window`` /
  ``EntityWindows`` type aliases. Clean dependency tree:

  kinds → windows → scope → queries

``__init__.py`` re-exports preserve the existing test imports.
Smoke-tested inside the container: the same function object is
reachable via ``superset.versioning.activity._intersect_windows`` and
via ``superset.versioning.activity.windows._intersect_windows``.

Surfaced by clean-code-review v2 (G6).
