+++ superset/versioning/activity/scope.py
@@ -14,37 +14,23 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Window arithmetic and scope resolution.
+"""Scope resolution — turn a path entity into the related-entity walk.
 
-The activity-view fetches change records across an entity's transitive
-dependency chain, time-bounded by when each relationship was active.
-This module collects all the pure functions that build the
-``list[EntityWindows]`` scope passed to
-:func:`~superset.versioning.activity.queries._fetch_change_records`:
+Composes :mod:`~superset.versioning.activity.queries` (Phase A
+relationship walks) and :mod:`~superset.versioning.activity.windows`
+(pure interval arithmetic) into the
+``list[EntityWindows]`` scope that
+:func:`~superset.versioning.activity.queries._fetch_change_records`
+consumes.
 
-* :func:`_intersect_windows` / :func:`_union_windows` — pure interval
-  arithmetic on half-open ``[start_tx, end_tx)`` ranges.
-* :func:`_row_within_any_window` — Python post-filter for records the
-  SQL fetch can't pre-narrow (used inside the orchestrator after the
-  per-kind fetch).
-* :func:`_merge_entity_windows` — collapses repeated entity entries
-  into one row per ``(api_kind, entity_id)`` with a minimal disjoint
-  cover of windows. Keeps the OR-clause count in
-  :func:`_fetch_change_records` proportional to *distinct* validity
-  intervals, not the number of shadow rows.
-* :func:`_resolve_scope` / :func:`_resolve_dashboard_scope` /
-  :func:`_resolve_chart_scope` / :func:`_resolve_related_scope` —
-  branch by path-kind to compute the full related-entity scope.
-
-The DB-touching relationship traversers used by the dashboard/chart
-scope resolvers (``_charts_attached_to_dashboard``,
-``_datasets_used_by_chart``, ``_batch_datasets_used_by_charts``) live
-next door in :mod:`~superset.versioning.activity.queries`.
+The functions here read the DB (via the Phase A helpers in
+:mod:`~superset.versioning.activity.queries`); the pure window-
+arithmetic functions previously colocated here now live in
+:mod:`~superset.versioning.activity.windows` so the package no longer
+needs a lazy import to dodge a ``scope ↔ queries`` cycle.
 """
 
 from __future__ import annotations
-
-from typing import Any
 
 from superset.versioning.activity.kinds import EntityWindows, Window
 from superset.versioning.activity.queries import (
@@ -52,40 +38,10 @@ from superset.versioning.activity.queries import (
     _charts_attached_to_dashboard,
     _datasets_used_by_chart,
 )
-
-
-def _intersect_windows(outer: Window, inner: Window) -> Window | None:
-    """Intersect two half-open ``[start_tx, end_tx)`` windows.
-
-    Returns the clipped overlap, or ``None`` when they are disjoint.
-    ``end_tx = None`` means "open ended (current)" and acts like
-    positive infinity.
-    """
-    o_start, o_end = outer
-    i_start, i_end = inner
-    start = max(o_start, i_start)
-    end: int | None
-    if o_end is None:
-        end = i_end
-    elif i_end is None:
-        end = o_end
-    else:
-        end = min(o_end, i_end)
-    if end is not None and end <= start:
-        return None
-    return (start, end)
-
-
-def _row_within_any_window(row: dict[str, Any], windows: list[Window]) -> bool:
-    """``True`` iff ``row['transaction_id']`` falls inside at least one
-    of *windows*. Half-open interval semantics match
-    :func:`_intersect_windows`."""
-    if not windows:
-        return False
-    tx_id = row["transaction_id"]
-    return any(
-        start <= tx_id and (end is None or tx_id < end) for start, end in windows
-    )
+from superset.versioning.activity.windows import (
+    _intersect_windows,
+    _merge_entity_windows,
+)
 
 
 def _resolve_scope(path_kind: str, path_id: int, include: str) -> list[EntityWindows]:
@@ -145,53 +101,3 @@ def _resolve_chart_scope(slice_id: int) -> list[EntityWindows]:
     for dataset_id, window in _datasets_used_by_chart(slice_id):
         scope.append(("SqlaTable", dataset_id, [window]))
     return _merge_entity_windows(scope)
-
-
-def _merge_entity_windows(scope: list[EntityWindows]) -> list[EntityWindows]:
-    """Collapse repeated ``(api_kind, entity_id)`` entries by unioning
-    their window lists, and collapse overlapping/touching windows
-    within each entity into one.
-
-    The OR-clause in
-    :func:`~superset.versioning.activity.queries._fetch_change_records`
-    generates one branch per (kind, id, window) tuple. Without the
-    within-entity union, a chart that's been attached-and-detached
-    many times (or that repeated fixture loads have populated the M2M
-    shadow for) yields a separate clause per redundant window — at
-    ~10 entities × ~50 windows the SQL hits SQLite's
-    ``SQLITE_MAX_EXPR_DEPTH`` (1000). Merging here keeps the clause
-    count proportional to the number of *distinct* validity intervals,
-    not the number of shadow rows.
-    """
-    merged: dict[tuple[str, int], list[Window]] = {}
-    for api_kind, entity_id, windows in scope:
-        merged.setdefault((api_kind, entity_id), []).extend(windows)
-    return [
-        (api_kind, entity_id, _union_windows(windows))
-        for (api_kind, entity_id), windows in merged.items()
-    ]
-
-
-def _union_windows(windows: list[Window]) -> list[Window]:
-    """Sort + merge overlapping/touching half-open intervals.
-
-    Pure function — no DB. Touching ``[a, b)`` and ``[b, c)`` merge into
-    ``[a, c)``. ``end_tx = None`` (open-ended) absorbs everything to its
-    right. Returns a minimal disjoint cover of the input set.
-    """
-    if not windows:
-        return []
-    sorted_windows = sorted(windows, key=lambda w: w[0])
-    out: list[Window] = [sorted_windows[0]]
-    for start, end in sorted_windows[1:]:
-        prev_start, prev_end = out[-1]
-        if prev_end is None:
-            # Prior window is open-ended; it absorbs everything past.
-            continue
-        if start <= prev_end:
-            # Overlapping or touching — extend the prior window.
-            new_end: int | None = None if end is None else max(prev_end, end)
-            out[-1] = (prev_start, new_end)
-        else:
-            out.append((start, end))
-    return out
